concluido

+++ app/validation/check_posts.php
@@ -62,19 +62,22 @@ function show_comments() {
     global $table_comments, $table_users, $table_posts, $autor, $title, $text, $id, $erro_texto;
     $conn = connect_db();
     $text = $_GET['text'];
-    $sql = "SELECT A.*, C.name FROM $table_comments A LEFT JOIN $table_posts B 
-            ON A.post_id = B.id LEFT JOIN $table_users C ON C.id = A.user_id";      
-    $result = mysqli_query($conn, $sql); // First parameter is just return of "mysqli_connect()" function
+    $sql = "SELECT A.id AS 'ID', A.user_id, A.content, B.id,C.name FROM $table_comments A, $table_posts B, $table_users C 
+            WHERE B.id = $id AND A.post_id = $id AND C.id = A.user_id";      
+    $result = mysqli_query($conn, $sql);
+     
    // echo "<div class='container'>";
-    while ($row = mysqli_fetch_assoc($result)) { // Important line !!! Check summary get row on array ..
-        $cid = $row['id'];
+    while ($row = mysqli_fetch_assoc($result)) { 
+        $cid = $row['ID'];
+        $user_id = $row['user_id'];
         echo "<h3>".$row['name']."</h3>";
         $comment = $row['content'];
-        if (isset($_GET['edit'])&&($cid == $_GET['edit'])) {
+
+        if (isset($_GET['edit'])&&($cid == $_GET['edit'])){ 
             echo "<form action='' enctype='multipart/form-data' id='form-edit' method='POST'>";
             echo '<div class="form-group <?php if(!empty($erro_texto)){echo "has-error";}"?>';
             echo "<label for='inputC' class='col-sm-3 control-label'>Editar Comentario</label>";
-            echo "<textarea required type='text' style='padding-right: 300px;' rows='5' name='ec' value='$comment'></textarea><br>";
+            echo "<textarea required type='text' style='padding-right: 250px;' rows='5' name='ec' value='$comment'></textarea><br>";
             echo "<div id='erro-texto'></div>";
             echo "<?php if (!empty($erro_texto)): ?>
                     <span class='help-block'><?php echo $erro_texto ?></span>
@@ -85,14 +88,17 @@ function show_comments() {
             echo "</form>";
            
         }
-        else { echo "<p>$comment</p>";
+        else  { 
+            echo "<div style='word-wrap: break-word; width: 800px;'><p>$comment</p></div>";
             if ($row['name'] == $autor) {
-                echo "<a href='view.php?id=$id&title=$title&text=$text&edit=$cid'>Editar</a>"." ";
-                echo "<a href='validation/comment_delete.php?id=$cid'>Excluir</a></p>";
+                if ($_SESSION['user_id'] == $user_id) {
+                    echo "<a href='view.php?id=$id&title=$title&text=$text&edit=$cid'>Editar</a>"." ";
+                    echo "<a href='validation/comment_delete.php?id=$cid'>Excluir</a></p>";
+                }
             }
         }
     }
-    //echo "</div>";
+   
     disconnect_db($conn);
 }
